checkpoint - moving functions into spawner package
